:truck: move CampaignParticipationOverview to prescription
