+++ api/tests/prescription/campaign-participation/unit/domain/read-models/CampaignParticipationOverview_test.js
@@ -1,8 +1,11 @@
 import dayjs from 'dayjs';
 
-import { CampaignParticipationStatuses, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
-import { CampaignParticipationOverview } from '../../../../src/shared/domain/read-models/CampaignParticipationOverview.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { CampaignParticipationOverview } from '../../../../../../src/prescription/campaign-participation/domain/read-models/CampaignParticipationOverview.js';
+import {
+  CampaignParticipationStatuses,
+  CampaignTypes,
+} from '../../../../../../src/prescription/shared/domain/constants.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 const { SHARED, STARTED } = CampaignParticipationStatuses;
